backend/requirements.txt
# SoundWars Flask API - Dependencies

# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0

# Database
PyMySQL==1.1.0
cryptography==41.0.7
SQLAlchemy==2.0.23

# Security
Werkzeug==3.0.1
bcrypt==4.1.2
bleach==6.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0

# For AWS deployment (optional)
gunicorn==21.2.0
boto3==1.34.0
